2020/04: List passport fields in rules array only
The set of required passport fields is written out four times: once to
define the array of strict validation rules, once during lax validation,
and twice during strict validation.  This is fragile and un-DRY (WET?).

Rewrite the validation algorithms to iterate over the rules array, so
the passport fields are specified in just one place.