Delete puzzle text
The puzzle text is unambiguously copyrighted material, and the website
explicitly states [*] that "[the] design elements, language, styles, and
concept of Advent of Code are all the sole property of Advent of Code
and may not be replicated or used by any other person or entity without
express written consent of Advent of Code."

One could make a case that including the puzzle text in a repository of
solutions to said puzzles qualifies as fair use, but I don't care enough
to try; I only included the text for my own convenience.

I'm letting the input stay.  I'm not sure it's actually copyrightable,
and in any case it's necessary to run the solutions, so the fair use
argument might be stronger than for the puzzle text.

Ignore local copies of the puzzle text that I'll keep around for myself.

  [*] https://adventofcode.com/about